Fix creation of RMF files in Linux package.
The combination of new Boost (specifically, the container::flat_map
and container::flat_set types) and the old gcc 4.1 resulted in RMF
generating corrupt output files, missing some of the node attributes.
Amongst other things, this caused the tutorial to fail. Fix this by
building with the newer gcc 4.4 instead, and add a test to make sure
that we can read back in RMF files that we create.