Build for Python 2.7 on Windows with Anaconda's boost.
Modern Anaconda on Windows requires that we build using
the same version of Visual Studio that was used to build
Python. In the case of Python 2.7, this is Visual Studio
2008 (vc9). Patch IMP to build with VS 2008 and require
vc9-enabled dependencies.

Note that we use stdint.h, which is not provided by
VS 2008, so we include a suitable substitute with this
recipe.