Build against newer OpenCV on Linux.
It turned out the older OpenCV we were previously
linking against was no longer updated by the
Anaconda folks, and so pulled in an older version
of numpy. This caused a conflict when trying to use
IMP at the same time as other packages that used
the newest numpy (such as h5py).