Update to latest DSSP and dependencies
Update to the latest upstream version (3.0.0) of DSSP,
and rebuild against the latest version of Boost. Pin all
dependencies so things don't break next time zlib or bzip2
are updated. (Note that this won't currently work on Mac
since DSSP 3 needs C++0x, but that conflicts with the conda
Boost package.) Closes #3.