Require libboost, not boost, at runtime
Requiring boost doesn't work for users that have the
conda-forge channel, since conda may choose the
conda-forge 'boost' package, which isn't compiled with
C++11, so isn't compatible with our packages (which are).
libboost is only present in the conda default channel,
not conda-forge, so the dependency is no longer ambiguous.
Closes #4.