Work around VS2008 internal compiler error
Visual Studio 2008 fails to build the IMP.isd
and IMP.pmi Python wrappers with an internal
compiler error. Since we need to build Python
extensions on Windows with the same version of
VS that the Python folks used, we don't have the
option of using a newer VS that fixes this problem.
Instead, slightly modify the affected code to avoid
the compiler error.