Build CGAL 4.12 for Windows with Python 2.7
We can't use the latest CGAL (4.14) on Windows
with Python 2.7, since we have to build with
Visual Studio 2008, which doesn't understand some
of the newer C++ syntax used by CGAL 4.14. But we
can't keep using CGAL 4.10 either, since it is
built against an older Boost which is no longer
available, and its CMake config is considered
invalid by newer CMake. Build 4.12 against the
latest Boost.