Fix Python search path
The standalone installer includes some logic to make
it work with multiple Python versions, which isn't
necessary with Anaconda since the Anaconda environment
handles that already. It's a harmless addition for Python
up to 3.7 as it just adds a (non-existent) directory to
the search path, but Python 3.8 changed the way the path
was handled, and reports an error if a path element
doesn't exist.

Since this logic isn't needed, just chop it out.